delete some moved stuff
