+++ tools/cli/flags.go
@@ -220,6 +220,7 @@ const (
 	FlagLowerShardBound                   = "lower_shard_bound"
 	FlagUpperShardBound                   = "upper_shard_bound"
 	FlagInputDirectory                    = "input_directory"
+	FlagDBInvariantChecks                 = "db_invariant_checks"
 )
 
 var flagsForExecution = []cli.Flag{
